// jimi-youping/src/pages/supplier/snapshot/data.d.ts
/**
 * 列表项
 * @param {id} 商品ID
 * @param {name} 商品名称
 * @param {sku} 商品sku
 * @param {thumb_url} 草稿图片
 * @param {cate_id} 草稿分类
 * @param {supply_price} 供应商价
 * @param {sort} 排序
 * @param {stock} 库存
 * @param {on_shelf} 上下架状态（0下架，1上架）
 * @param {cate_id} 分类ID
 * @param {pending_confirm_count} 待确认/待扣除数量
 * @param {pending_deliver_count} 待发货数量
 * @param {cate_info} 分类信息
 * @param {supplier_info} 分类信息
 */
export type TableListItem = {
  id: number;
  name: string;
  sku: string;
  thumb_url: string;
  cate_id: string | number;
  supply_price: string;
  sort: number;
  stock: number;
  on_shelf: number;
  pending_confirm_count: string | number;
  pending_deliver_count: string | number;
  cate_info: Record<string, any[]>;
  supplier_info: {
    id: string | number;
    name: string;
  };
};

/**
 * 列表param
 * @param {current} 当前页数（默认为1）
 * @param {pageSize} 每页显示条数（默认为10）
 * @param {name} 检索关键字
 * @param {supplier_id} 上级ID
 * @param {cate_id} 分类ID
 * @param {allow_oversell} 是否超售
 * @param {start} 开始时间
 * @param {end} 结束时间
 * @param {on_shelf} 状态（0下架1上架，空不进行该字段检索）
 * @returns
 */
export type TableListParams = {
  current?: number;
  pageSize?: number;
  name?: string;
  supplier_id?: string | number;
  cate_id?: string | number;
  allow_oversell?: string | number;
  on_shelf?: string | number;
  start?: string | number;
  end?: string | number;
  sorter?: any;
  filter?: any;
};